Move ServerInterface to separate header
